requirements.txt
asgiref==3.8.1
captcha==0.6.0
channels==4.1.0
crispy-bootstrap4==2024.10
Django==4.2.16
django-admin-filter==0.7
django-bootstrap-form==3.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-ckeditor5==0.0.4
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-extra-views==0.15.0
django-filter==24.3
django-js-asset==2.2.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-taggit==6.1.0
djangorestframework==3.15.2
Faker==33.0.0
lorem==0.1.1
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
